refactor(ee): move multi-tenant sso to /web/src/ee
