+++ web/src/ee/features/multi-tenant-sso/utils.ts
@@ -5,12 +5,10 @@ import OktaProvider from "next-auth/providers/okta";
 import CognitoProvider from "next-auth/providers/cognito";
 import Auth0Provider from "next-auth/providers/auth0";
 import AzureADProvider from "next-auth/providers/azure-ad";
-import { isEeAvailable } from "..";
+import { isEeEnabled } from "@/src/ee/utils/isEeEnabled";
 import { type SsoConfig, prisma } from "@langfuse/shared/src/db";
-import { encrypt, decrypt } from "@langfuse/shared/encryption";
+import { decrypt } from "@langfuse/shared/encryption";
 import { SsoProviderSchema } from "./types";
-import { type NextApiRequest, type NextApiResponse } from "next";
-import { env } from "../env";
 import { CustomSSOProvider } from "@langfuse/shared/src/server";
 
 // Local cache for SSO configurations
@@ -25,7 +23,7 @@ let cachedSsoConfigs: {
  * @returns {Promise<SsoProviderSchema[]>} - A list of all SSO configurations. Empty array if none are configured or EE is not available.
  */
 async function getSsoConfigs(): Promise<SsoProviderSchema[]> {
-  if (!isEeAvailable) return [];
+  if (!isEeEnabled) return [];
   const CACHE_TTL = 60 * 1000; // 1 minute
   const DB_MAX_WAIT = 2000; // 2 seconds
   const DB_TIMEOUT = 3000; // 3 seconds
@@ -43,7 +41,7 @@ async function getSsoConfigs(): Promise<SsoProviderSchema[]> {
         {
           maxWait: DB_MAX_WAIT,
           timeout: DB_TIMEOUT,
-        }
+        },
       );
     } catch (e) {
       // cache empty array to prevent repeated DB calls on error
@@ -65,7 +63,7 @@ async function getSsoConfigs(): Promise<SsoProviderSchema[]> {
         } catch (e) {
           console.error(
             `Failed to parse SSO provider config for domain ${v.domain}`,
-            e
+            e,
           );
           return null;
         }
@@ -87,7 +85,7 @@ async function getSsoConfigs(): Promise<SsoProviderSchema[]> {
  * @returns {Promise<Provider[]>} - A list of all custom SSO providers.
  */
 export async function loadSsoProviders(): Promise<Provider[]> {
-  if (!isEeAvailable) return [];
+  if (!isEeEnabled) return [];
 
   const ssoConfigs = await getSsoConfigs();
 
@@ -105,7 +103,7 @@ export async function loadSsoProviders(): Promise<Provider[]> {
  * @returns `true` if any custom SSO provider is configured in the database.
  */
 export async function isAnySsoConfigured(): Promise<boolean> {
-  if (!isEeAvailable) return false;
+  if (!isEeEnabled) return false;
   const ssoConfigs = await getSsoConfigs();
   return ssoConfigs.length > 0;
 }
@@ -117,11 +115,11 @@ export async function isAnySsoConfigured(): Promise<boolean> {
  * @returns `providerId` or null if none is configured or EE is not available.
  */
 export async function getSsoAuthProviderIdForDomain(
-  domain: string
+  domain: string,
 ): Promise<string | null> {
-  if (!isEeAvailable) return null;
+  if (!isEeEnabled) return null;
   const ssoConfig = (await getSsoConfigs()).find(
-    (ssoConfig) => ssoConfig.domain === domain.toLowerCase()
+    (ssoConfig) => ssoConfig.domain === domain.toLowerCase(),
   );
 
   if (!ssoConfig) return null;
@@ -188,7 +186,7 @@ const dbToNextAuthProvider = (provider: SsoProviderSchema): Provider | null => {
     // eslint-disable-next-line no-unused-vars
     const _: never = provider;
     throw new Error(
-      `Unrecognized SSO provider for domain ${(provider as any).domain}`
+      `Unrecognized SSO provider for domain ${(provider as any).domain}`,
     );
   }
 };
@@ -200,76 +198,8 @@ const dbToNextAuthProvider = (provider: SsoProviderSchema): Provider | null => {
  * @returns {string} - The providerId used in NextAuth.
  */
 const getAuthProviderIdForSsoConfig = (
-  dbSsoConfig: SsoProviderSchema
+  dbSsoConfig: SsoProviderSchema,
 ): string => {
   if (!dbSsoConfig.authConfig) return dbSsoConfig.authProvider;
   return `${dbSsoConfig.domain}.${dbSsoConfig.authProvider}`;
 };
-
-export async function createNewSsoConfigHandler(
-  req: NextApiRequest,
-  res: NextApiResponse
-) {
-  try {
-    if (!isEeAvailable) {
-      res.status(403).json({ error: "EE is not available" });
-      return;
-    }
-    // allow only POST requests
-    if (req.method !== "POST") {
-      res.status(405).json({ error: "Method Not Allowed" });
-      return;
-    }
-    // check if ADMIN_API_KEY is set
-    if (!env.ADMIN_API_KEY) {
-      res.status(500).json({ error: "ADMIN_API_KEY is not set" });
-      return;
-    }
-    if (!env.ENCRYPTION_KEY) {
-      res.status(500).json({ error: "ENCRYPTION_KEY is not set" });
-      return;
-    }
-    // check bearer token
-    const { authorization } = req.headers;
-    if (!authorization) {
-      res
-        .status(401)
-        .json({ error: "Unauthorized: No authorization header provided" });
-      return;
-    }
-    const [scheme, token] = authorization.split(" ");
-    if (scheme !== "Bearer" || !token || token !== env.ADMIN_API_KEY) {
-      res.status(401).json({ error: "Unauthorized: Invalid token" });
-      return;
-    }
-
-    const body = SsoProviderSchema.safeParse(req.body);
-    if (!body.success) {
-      res.status(400).json({ error: body.error });
-      return;
-    }
-
-    const { domain, authProvider, authConfig } = body.data;
-
-    const encryptedClientSecret = authConfig
-      ? {
-          ...authConfig,
-          clientSecret: encrypt(authConfig.clientSecret),
-        }
-      : undefined;
-
-    await prisma.ssoConfig.create({
-      data: {
-        domain,
-        authProvider,
-        authConfig: encryptedClientSecret,
-      },
-    });
-    res.status(201).json({
-      message: "SSO configuration created successfully",
-    });
-  } catch (e) {
-    console.log(e);
-    res.status(500).json({ error: "Internal Server Error" });
-  }
-}
